+++ core/src/fr/leroideskiwis/mapgame/specialobjects/HorizontalOpenPath.java
@@ -37,7 +37,7 @@ public class HorizontalOpenPath extends SpecialObj {
     }
 
     @Override
-    public double chance() {
-        return 0.075;
+    public float chance() {
+        return 0.075f;
     }
 }
