+++ core/src/fr/leroideskiwis/mapgame/specialobjects/Respawn.java
@@ -21,7 +21,7 @@ public class Respawn extends SpecialObj {
     }
 
     @Override
-    public double chance() {
-        return 0.06;
+    public float chance() {
+        return 0.06f;
     }
 }
